Moved and renamed Navbar.jsx from components/users/Navbar.jsx --> components/NavBar.jsx. Uncommented NavBar for now and imported NavBar in App.js
